requirements.txt
# VOXCODE v2.0 Dependencies
# Voice-Controlled Windows Automation with Intelligent Agent

# ==================== Core ====================
requests>=2.28.0
python-dotenv>=1.0.0

# ==================== Terminal UI ====================
textual>=0.47.0
rich>=13.0.0

# ==================== Windows Automation ====================
pyautogui>=0.9.54
pywinauto>=0.6.8
Pillow>=10.0.0

# ==================== Vision / OCR ====================
# EasyOCR for text detection
easyocr>=1.7.0

# OmniParser dependencies (for advanced UI detection)
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
supervision>=0.18.0
opencv-python>=4.8.0
huggingface_hub>=0.20.0

# Optional: Icon captioning (Florence-2)
# transformers>=4.36.0
# accelerate>=0.25.0

# ==================== Voice ====================
# STT (Speech-to-Text) - Whisper
openai-whisper>=20231117
pyaudio>=0.2.14
numpy>=1.24.0

# TTS (Text-to-Speech)
pyttsx3>=2.90

# Keyboard handling (global hotkeys)
keyboard>=0.13.5

# ==================== NEW: Browser Automation ====================
# Browser-Use: LLM-powered browser control via Playwright
# This enables reliable web automation (YouTube, Google, etc.)
browser-use>=0.1.40
playwright>=1.40.0

# Note: After installing, run: playwright install chromium

# ==================== NEW: Ollama VLM ====================
# For Vision-Language Model integration
# Install models with: ollama pull qwen2.5vl:7b

# ==================== Setup Instructions ====================
# 1. Install Python dependencies:
#    pip install -r requirements.txt
#
# 2. Install Ollama (for LLM/VLM):
#    https://ollama.com/download
#
# 3. Pull required models:
#    ollama pull qwen2.5:7b         # Text LLM for planning
#    ollama pull qwen2.5vl:7b       # Vision-Language Model (recommended)
#    OR
#    ollama pull minicpm-v:8b       # Alternative lighter VLM
#
# 4. Start Ollama:
#    ollama serve
#
# 5. Run VOXCODE:
#    python main.py
